Change the path to store multi-base-package PVEPs (#196)
* fix: use hash of .base instead of {base, parameterisation} for multi-package PVEPs

The parameterisation options of a synthesized package can change after its creation.
This means that the digest of the PVEP will evaluate to a different value. If we then
try to launch the synthesized PVEP we will end up giving elaunch.py a path to a directory
that does not exist. Therefore, we should create a unique identifier of the base packages for
a multi-base-package PVEP and use that to craft the path to the underlying virtual experiment
* fix: field options.generateParameterisation in /relationship/<identifier>/synthesize/<name>
* chore: Test whether the files of a multi-package PVEP exist right before launching it

Resolves https://github.ibm.com/st4sd/st4sd-runtime-service/issues/195

Signed-off-by: Vassilis Vassiliadis <vassilis.vassiliadis@ibm.com>