feat: optimize GET /instances/ (#208)
The optimization is to read the st4sd-runtime-service ConfigMap just once (to find where the PVEP database is).
Then find out all the unique PVEP identifiers that the instances are associated with. From there build a
mapping of PVEP-identifier to PVEP dictionary by querying the PVEP database. Finally use this cache of
PVEPs to augment the list of instances.

Resolves https://github.ibm.com/st4sd/st4sd-runtime-service/issues/154

Signed-off-by: Vassilis Vassiliadis <vassilis.vassiliadis@ibm.com>